Changed the schema so that now features are encapsulated within layers (as the shapefile defines it)
This allows shapefiles to have multiple layers with different featuresets -- previously, all of these features
would be queryable with allGeoFeature and the differentiation of layers could be done with layerName

Additionally, data from different files now shows up as a different node in graphQL

Previously, all layers from different files would be queryable under allGeoLayer, which could be differentiated
by their names. Now, each file will show up as all[fileName]Layer.

Worth noting that this is a quick fix for the issue of points and polygons conflicting with the way
that they store coordinate data -- hopefully I will be able to implement a more robust fix for this,
as the issue can still exist if shapefiles have both points and polygons.